Updated the website with Proper Html5 syntax structure
Added content to three of my projects (though still in process). Used grids and flexbox to layout the projects in the portfolio section and navigation list respectively.

I also included media queries to make the site responsive and modified the header, content in the Css Style sheet.

The site is still under maintaince.